Frameworks: ElectronBeam get some love (1/2)
This commit removes the CRT On animation because it is just
not nice implemented from AOSP and not really needed. If ppl
turn on the screen they want to see fast their content. But still
CRT animation needed some love. So this commit introduces another
SlimBean mode. Crt animation mode:

Features:
- Crt Off animation controll (on/off like it was before)
- You can select different crt off animation modes
- always horizontal (stock behaviour)
- always vertical (screen collapse from left and right to the middle)
- orientation dependent (means portrait = horizontal animation, landscape = vertical animation)

cheers

Change-Id: I193313619e8b26f7c4a08e864310d4eaa62225f3